# swagger/swagger.json
  {
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "shopmart",
      "contact": {}
    },
    "host": "shopmart-nodeapp.herokuapp.com",
    "basePath": "/api",
    "securityDefinitions": {},
    "schemes": [
      "https","http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/user/changePassword": {
        "post": {
          "summary": "change password",
          "tags": [
            "user"
          ],
          "operationId": "changepassword",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/changepasswordrequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      }
    },
    "definitions": {
...
      "changepasswordrequest": {
        "title": "changepasswordrequest",
        "example": {
          "email": "suhagiyadivyesh1515@gmail.com",
          "oldPassword": "divyesh@111",
          "newPassword": "divyesh@new",
          "confirmPassword": "divyesh@new"
        },
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "oldPassword",
          "newPassword",
          "confirmPassword"
        ]
      }
    },
...
  }